RequirementDefinition modified event flow refactor
